requirements.txt
colorama
python-dotenv
discord.py
pandas